08-17 19:32

+++ homework/src/main/java/com/example/homework/Domain/entity/OrderLine.java
@@ -20,5 +20,5 @@ public class OrderLine {
     @TableField("price")
     private BigDecimal price;
     @TableField("quantity")
-    private Double quantity;
+    private BigDecimal quantity;
 }
